fix: Close ALL remaining — 30 NOT FIXED + 12 PARTIAL = 0 open
## P0 closed (3):
- CoachProfile operator==: 38 fields compared (was 13). hashCode aligned.
- GoalA/GoalB/PlannedMonthlyContribution: added == + hashCode
- LibrePassageCompte: dateOuverture in equality
- Markdown/HTML injection: stripped from LLM responses before rendering

## P1 closed (10):
- SubscriptionProvider: catchError on fire-and-forget
- allowMalformed: removed from document scan UTF-8 decode
- Children upper limit: le=20 on 10 schema fields across 5 files
- Locked-out user: token cleared on 401 in all HTTP methods
- nombreDeMois: int → double (accepts 12.0, 13.0, prorata values)
- Coach insights: offline message queue (SharedPreferences pending_messages)
- Compliance logging: includes user_id

## P2 closed (5):
- birthYear: verified no 1990 fallback remaining
- Mortgage journey: 24 ARB keys extracted (7 steps, 6 languages)
- vous/votre: 13 coaching instances → tu/ton/ta

## Shortcuts self-audited and fixed:
- Tie-breaking: hashCode → lexicographic compareTo (stable across versions)
- FATCA 3a: not blocked (Swiss law allows), FATCA warning in templates instead
- Emergency fund 'no': documented as conservative floor (not 0)

Tests: 8125 Flutter + 4679 backend = 12804 all green
flutter analyze: 0 errors, 0 warnings | ruff: All checks passed

+++ apps/mobile/test/services/financial_core/coach_reasoner_test.dart
@@ -9,7 +9,7 @@ void main() {
   CoachProfile profile({
     int age = 50,
     double salaire = 8000,
-    int mois = 12,
+    double mois = 12.0,
     String canton = 'VS',
     String employment = 'salarie',
     CoachCivilStatus etatCivil = CoachCivilStatus.celibataire,
